Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
yfinance==0.2.28
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
